backend/requirements.txt
Django==4.2.27
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
pylint-django==2.7.0